Move UniqueIdMixin to the mixins folder
